Add files via upload
I have copied and pasted my code into this word document because every time I tried to open the stock excel file it shows as "This file is corrupted and cannot be opened". I have also tried to take screenshots in order to show prove of it. However, I have been researching for a plenty of ways to take screenshot in a windows computer and nothing has worked. 

This has been my last resort. I will make sure to communicate this with the professor and TAs.